third_party/blink/web_tests/fast/css/invalidation/input-search-incremental.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
    input + div { color: pink }
</style>
<input id="searchInput" type="search"></input>
<div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
description("Setting the incremental attribute on an input type=search")

shouldBeDefined(window.internals);

searchInput.offsetTop;
getComputedStyle(searchInput); // Force recalc.
searchInput.setAttribute("incremental", "");

shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "0");

</script>
