third_party/blink/web_tests/fast/css/invalidation/input-search-incremental-expected.txt

Setting the incremental attribute on an input type=search

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Internals] is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
